requirements.txt
pytorch
numpy
matplotlib
gymnasium
robosuite
dm-control
h5py
seaborn